src/third_party/blink/web_tests/media/track/track-css-user-settings-override-internal-settings.html

<!DOCTYPE html>
<title>Test that WebVTT objects are being styled correctly based on user settings that should override default settings.</title>
<script src="../media-controls.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<video>
    <track src="captions-webvtt/styling.vtt" kind="captions" default>
</video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = "../content/test.ogv";

    video.oncanplaythrough = t.step_func_done(function() {
        var cue = textTrackCueElementByIndex(video, 0).firstChild;
        var displayTree = textTrackCueElementByIndex(video, 0);
        var cueStyle = getComputedStyle(cue);
        var displayTreeStyle = getComputedStyle(displayTree);
        // These are the expected default cue settings per spec
        // http://dev.w3.org/html5/webvtt/#applying-css-properties-to-webvtt-node-objects
        assert_equals(cueStyle.color, "rgb(255, 255, 255)");
        assert_equals(cueStyle.backgroundColor, "rgba(0, 0, 0, 0.8)");
        assert_equals(cueStyle.fontFamily, "sans-serif");
        assert_equals(displayTreeStyle.backgroundColor, "rgba(0, 0, 0, 0)");
        assert_equals(displayTreeStyle.padding, "0px");

        // Apply user settings for color and font-size and verify that the other internal settings are retained.
        internals.settings.setTextTrackTextColor("purple");
        internals.settings.setTextTrackTextSize("14px");
        internals.settings.setTextTrackWindowColor("rgba(0, 0, 0, 0.8)");
        internals.settings.setTextTrackWindowPadding("5px");

        video.currentTime = 0.3;

        var cue = textTrackCueElementByIndex(video, 0).firstChild;
        var displayTree = textTrackCueElementByIndex(video, 0);
        var cueStyle = getComputedStyle(cue);
        var displayTreeStyle = getComputedStyle(displayTree);
        assert_equals(cueStyle.color, "rgb(128, 0, 128)");
        assert_equals(cueStyle.fontSize, "14px");
        assert_equals(displayTreeStyle.backgroundColor, "rgba(0, 0, 0, 0.8)");
        assert_equals(displayTreeStyle.padding, "5px");
        // When there is no user setting specified for background-color and font-family, the internal settings are applied.
        assert_equals(cueStyle.backgroundColor, "rgba(0, 0, 0, 0.8)");
        assert_equals(cueStyle.fontFamily, "sans-serif");
    });
});
</script>
